# Speed up tensor.storage_offset