--- a/aten/src/ATen/native/native_functions.yaml
+++ b/aten/src/ATen/native/native_functions.yaml
@@ -2294,10 +2294,6 @@
 
 # wrappers for legacy TH methods
 
-- func: storage_offset(Tensor self) -> int64_t
-  variants: method
-  device_guard: false
-
 - func: ndimension(Tensor self) -> int64_t
   variants: method
   device_guard: false